+++ tests/JsonApi/Transformer/AbstractDocumentTest.php
@@ -1,0 +1,48 @@
+<?php
+namespace WoohooLabsTest\Yin\JsonApi\Transformer;
+
+use PHPUnit_Framework_TestCase;
+use WoohooLabs\Yin\JsonApi\Schema\JsonApi;
+use WoohooLabs\Yin\JsonApi\Schema\Links;
+use WoohooLabs\Yin\JsonApi\Transformer\AbstractDocument;
+use Zend\Diactoros\Response;
+
+class AbstractDocumentTest extends PHPUnit_Framework_TestCase
+{
+    public function testGetExtensions()
+    {
+        $document = $this->createDocument();
+        $this->assertEquals([], $document->getSupportedExtensions());
+    }
+
+    public function testGetRequiredExtensions()
+    {
+        $document = $this->createDocument();
+        $this->assertEquals([], $document->getSupportedExtensions());
+    }
+
+    /**
+     * @param \WoohooLabs\Yin\JsonApi\Schema\JsonApi|null $jsonApi
+     * @param array $meta
+     * @param \WoohooLabs\Yin\JsonApi\Schema\Links|null $links
+     * @return \WoohooLabs\Yin\JsonApi\Transformer\AbstractDocument
+     */
+    private function createDocument(JsonApi $jsonApi = null, array $meta = [], Links $links = null)
+    {
+        $mock = $this->getMockForAbstractClass(AbstractDocument::class);
+
+        $mock
+            ->method("getJsonApi")
+            ->willReturn($jsonApi);
+        $mock
+            ->method("getMeta")
+            ->withAnyParameters()
+            ->willReturn($meta);
+        $mock
+            ->method("getLinks")
+            ->withAnyParameters()
+            ->willReturn($links);
+
+        return $mock;
+    }
+}